+++ Service.API/Domain/Domain/BussinessObjects/FantasyBOs/LineupBO.cs
@@ -8,7 +8,7 @@ namespace Fantasy.API.Domain.BussinessObjects.FantasyBOs
 {
     public class LineupBO
     {
-        public int LineUpId { get; set; }
+        public long LineUpId { get; set; }
         public IEnumerable<PlayerBO> Players { get; set; }
         public UserBO User { get; set; }
     }
